Load Pyodide implementation from external capnproto file
This is quite work in progress. The idea is we can upload these binary capnproto
files to GCS and be able to select at runtime the capnproto module for the
version of Pyodide corresponding to the worker's compatibility date.

Currently I am just hardcoding the path. For some reason with `The requested
module 'pyodide:python-entrypoint-helper' does not provide an export named
'default'`. Adding logging info shows that the text and type of the module is
right so I'm not sure what is going wrong.